requirements.txt
pandas==1.5.1
Pillow==9.2.0
Pillow==10.0.0
psd_tools==1.9.23
pytz==2022.5
Requests==2.31.0
tkcalendar==1.6.1
